+++ src/collective/elastic/plone/api/services/kitsearch/configure.zcml
@@ -2,12 +2,12 @@
   xmlns="http://namespaces.zope.org/zope"
   xmlns:plone="http://namespaces.plone.org/plone">
 
-  <adapter factory=".get.Kitsearch" name="kitsearch"/>
+  <adapter factory=".post.Kitsearch" name="kitsearch"/>
 
   <plone:service
-    method="GET"
+    method="POST"
     for="zope.interface.Interface"
-    factory=".get.KitsearchGet"
+    factory=".post.KitsearchGet"
     name="@kitsearch"
     permission="zope2.View"
     />
