+++ src/collective/elastic/plone/api/services/kitsearch/post.py
@@ -1,0 +1,42 @@
+# -*- coding: utf-8 -*-
+from plone import api
+from plone.restapi.interfaces import IExpandableElement
+from plone.restapi.deserializer import json_body
+from plone.restapi.services import Service
+from zope.component import adapter
+from zope.interface import implementer
+from zope.interface import Interface
+
+
+@implementer(IExpandableElement)
+@adapter(Interface, Interface)
+class Kitsearch(object):
+    def __init__(self, context, request):
+        self.context = context.aq_explicit
+        self.request = request
+
+    def __call__(self, data):
+        result = {
+            "kitsearch": {
+                "@id": "{}/@kitsearch".format(
+                    self.context.absolute_url(),
+                ),
+            },
+        }
+        if not data:
+            return result
+
+        # TODO get items from elasticsearch
+        from collective.elastic.plone.api.services.kitsearch.response_example import (
+            elasticsearchresponse,
+        )
+
+        result["kitsearch"]["elasticsearchresponse"] = elasticsearchresponse
+        return result
+
+
+class KitsearchGet(Service):
+    def reply(self):
+        data = json_body(self.request)
+        service_factory = Kitsearch(self.context, self.request)
+        return service_factory(data=data)["kitsearch"]
